requirements.txt
# This is the `requirements.txt` icluding the necessary packages and libraries.
# You can test it yourself by running `pip install -r requirements.txt` in a new virtual environment.

# Requirements for brain response classification project
torch>=2.4.0
numpy>=2.2.0
Pillow>=11.0.0
scikit-learn>=1.6
torchvision>=0.15.0
